feat: split into many components, try to update children when user use arrow keyboard
