+++ packages/painperdu/package.json
@@ -18,6 +18,7 @@
     "@storybook/addon-links": "^7.6.3",
     "@storybook/addon-onboarding": "^1.0.9",
     "@storybook/blocks": "^7.6.3",
+    "@storybook/builder-vite": "^7.6.10",
     "@storybook/react": "^7.6.3",
     "@storybook/react-vite": "^7.6.3",
     "@storybook/test": "^7.6.3",
@@ -26,7 +27,8 @@
     "eslint-plugin-storybook": "^0.6.15",
     "storybook": "^7.6.3",
     "tailwindcss": "^3.3.6",
-    "ts-config-painperdu": "workspace:^"
+    "ts-config-painperdu": "workspace:^",
+    "vite": "^5.0.12"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.17",
