+++ packages/painperdu/src/components/PainPerduFooter/PainPerduFooter.tsx
@@ -1,4 +1,4 @@
-export const PainPerduFooter = () => {
+export const PainPerduFooter = (): JSX.Element => {
   return (
     <footer className="flex items-center bg-white rounded-md box-border h-11 justify-between	py-0 px-3 relative w-full z-30">
       <ul className="flex list-none	m-0 p-0">
