+++ packages/painperdu/src/hooks/use-command-manager.ts
@@ -1,0 +1,17 @@
+import type { CommandHandler } from '../types';
+
+interface Results {
+  shouldCallFn: boolean;
+};
+
+export const useCommandManager = (eventCode: string, commandHandler: CommandHandler ): Results => {
+	if (commandHandler[eventCode] === undefined) {
+	   return {
+			shouldCallFn: false,
+		};
+	};
+
+	return {
+	   shouldCallFn: true,
+	};
+}
