+++ packages/painperdu/src/types.ts
@@ -1,8 +1,12 @@
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-type PathItem = {
+export type PathItem = {
 	alias: string
 	path: string
 	isSelected: boolean
 }
 
-export type { PathItem }
+export type commandsAvailables = 'KeyK' | 'Escape' | 'ArrowDown' | 'ArrowUp';
+
+export type CommandHandler = {
+  [key in commandsAvailables]?: (args?: unknown) => void;
+}
